# public/swagger/dist/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Book-API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost"
    }
  ],
  "paths": {
...
    "/api/auth/refresh-tokens": {
      "post": {
        "summary": "Method to refresh tokens",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "JWT": []
          }
        ],
        "operationId": "refresh",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 | OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "HTTP-message": {
                      "type": "string",
                      "example": "200 | OK"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user_id": {
                          "type": "integer",
                          "example": 33
                        },
                        "access_token": {
                          "type": "string",
                          "example": "token"
                        },
                        "refresh_token": {
                          "type": "string",
                          "example": "token"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "401 | Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}